updated views configuration
