+++ healthApp/views/__init__.py
@@ -1,0 +1,4 @@
+from .admin import *
+from .auth import *
+from .main import *
+from .public import *
