+++ healthApp/views/public.py
@@ -1,0 +1,16 @@
+from django.shortcuts import render, redirect
+from django.contrib import messages
+from healthApp.models import *
+
+
+def index(request):
+    if 'user_id'  not in request.session:
+        return render(request, 'welcome.html')
+    else:
+        user = User.objects.get(id=request.session['user_id'])
+        logs = Log.objects.all().order_by('-updatedAt')
+        context = {
+            'user': user,
+            'logs': logs,
+        }
+        return render(request, 'index.html', context)
