+++ requirements.txt
@@ -2,6 +2,6 @@ bcrypt==3.2.0
 Django==3.2.8
 django-environ==0.8.1
 ipython==7.28.0
-mysql-connector-python 8.0.28
+mysql-connector-python==8.0.29
 mysqlclient==2.1.0
 Pillow==8.4.0
